+++ CHANGELOG.md
@@ -1,28 +1,52 @@
-## 1.0.0 (2015-09-21)
+## 0.7.0 (2015-09-21)
 
-Adding new functions :
+Adding renameKeys function in object module.
+
+## 0.6.0 (2015-09-21)
+
+Adding new modules & functions :
 
 Crypto Module : 
 - randomizedPassword : generate a random password
 - encrypt : encrypt given string
 - decrypt : decrypt a value encrypted with encrypt function
 
+## 0.5.0 (2015-09-21)
+
+Adding new modules & functions :
+
 Date Module : 
 
 - generateList : Generate a list of date between two given index
 
+## 0.4.0 (2015-09-21)
+
+Adding new modules & functions :
+
 Media Module : 
 
 - isValidImageFormat : Test is given image format is valid
+
+## 0.3.0 (2015-09-21)
+
+Adding new modules & functions :
 
 Object Module : 
 
 - renameKey : rename a given key for another given
 - inpect : get the current object to a string representation with a full depth
 
+## 0.2.0 (2015-09-21)
+
+Adding new modules & functions :
+
 Request Module : 
 
 - getHost : Get current host name from request object
+
+## 0.1.0 (2015-09-21)
+
+Adding new modules & functions :
 
 Strings Module : 
 - generateAsciiCharsList : Generate an list of chars from ascii table